requirements.txt
numpy~=2.2.1
opencv-python~=4.11.0.86
matplotlib~=3.10.1
colormath~=3.0.0
torch~=2.7.0
torchvision~=0.22.0
PyYAML~=6.0.2
tqdm~=4.67.1
scikit-learn~=1.6.1
pandas~=2.2.3
pillow~=11.1.0